add some unit tests and review structure a little
