+++ templates/html/test_template.html
@@ -1,0 +1,1 @@
+<html><header><title>Test Template. Please keep this file in this folder.</title></header><body>{{ .TestContentInjection }}</body></html>
