+++ templates/meta/test_template.json
@@ -1,0 +1,10 @@
+{
+    "name": "test_template",
+    "description": "email for testing purposes",
+    "templateFilename": "test_template.html",
+    "subject": "TestTemplateSubject",
+    "contentChunks":[
+        "TestContentInjection"
+    ],
+    "escapeTranslationChunks":[]
+}
