+++ templates/templates_test.go
@@ -1,0 +1,36 @@
+package templates
+
+import (
+	"testing"
+)
+
+const (
+	templatesPath = "."
+	templateName  = "test_template"
+)
+
+func Test_GetTemplateMeta(t *testing.T) {
+	var expectedTemplateSubject = "TestTemplateSubject"
+	// Get template Metadata
+	var templateMeta = getTemplateMeta(templatesPath + "/meta/" + templateName + ".json")
+
+	if templateMeta.Subject == "" {
+		t.Fatal("Template Meta cannot be found")
+	}
+
+	if templateMeta.Subject != expectedTemplateSubject {
+		t.Fatalf("Template Meta is found but subject \"%s\" not the one expected \"%s\"", templateMeta.Subject, expectedTemplateSubject)
+	}
+}
+
+func Test_GetBody(t *testing.T) {
+	// Get template Metadata
+	var templateMeta = getTemplateMeta(templatesPath + "/meta/" + templateName + ".json")
+	var templateFileName = templatesPath + "/html/" + templateMeta.TemplateFilename
+
+	var templateBody = getBody(templateFileName)
+
+	if templateBody == "" {
+		t.Fatalf("Template Body cannot be found: %s", templateFileName)
+	}
+}
